rearranged files and functions heavily
