+++ src/colors/colors_ops.c
