+++ src/init/init.c
@@ -1,17 +1,5 @@
 #include "minirt.h"
 
-/**
- * @brief Setup viewport 2D size
- * @todo FOV will eventually come from .rt file in submitted project
- */
-// void	init_vp_attributes(t_minirt *minirt)
-// {
-// 	minirt->vp->fov = M_PI / 2;
-// 	minirt->vp->aspect_ratio = (float)WIDTH / (float)HEIGHT;
-// 	minirt->vp->vp_h = 2.0f * tan(minirt->vp->fov / 2);
-// 	minirt->vp->vp_w = minirt->vp->aspect_ratio * minirt->vp->vp_h;
-// }
-
 void	init_minirt(t_minirt *minirt)
 {
 	minirt->world = NULL;
