+++ src/init/init_objects.c
@@ -1,0 +1,27 @@
+#include "minirt.h"
+
+void	init_objects(t_minirt *minirt)
+{
+	int			i;
+	t_list		*temp;
+	t_scene_obj	*obj;
+
+	temp = minirt->world->objects;
+	i = 0;
+	while(i < minirt->world->obj_count)
+	{
+		obj = (t_scene_obj *)temp->content;
+		obj->id = i;
+		if (obj->type == CYLINDER)
+		{
+			obj->min = -0.5;
+			obj->max = 0.5;
+			obj->closed = true;
+		}
+		obj->transform = generate_transformation_mtrx(minirt, obj);
+		obj->mat = init_material(minirt->world->amb_s.ratio);
+		color_convert(obj);
+		temp = temp->next;
+		i++;
+	}
+}
