+++ src/init/obj_rotation.c
