+++ src/lighting/lighting.c
