+++ src/raytracing/normal.c
